add validation

+++ src/main/java/lk/ijse/packageservice/api/PackageController.java
@@ -1,6 +1,8 @@
 package lk.ijse.packageservice.api;
 
 import lk.ijse.packageservice.dto.PackageDTO;
+import lk.ijse.packageservice.entity.Role;
+import lk.ijse.packageservice.exception.UnauthorizedException;
 import lk.ijse.packageservice.service.custom.PackageService;
 import lk.ijse.packageservice.util.ResponseUtil;
 import lombok.RequiredArgsConstructor;
@@ -8,9 +10,6 @@ import org.springframework.http.MediaType;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
 
-import java.text.ParseException;
-import java.text.SimpleDateFormat;
-import java.util.Date;
 import java.util.List;
 
 /**
@@ -29,80 +28,108 @@ public class PackageController {
     @GetMapping(value = "/getVehicles", params = {"page", "count", "category"}, produces = {MediaType.APPLICATION_JSON_VALUE})
     public ResponseUtil getVehiclesByCategory(@RequestParam Integer page,
                                               @RequestParam Integer count,
-                                              @RequestParam String category){
+                                              @RequestParam String category,
+                                              @RequestHeader("X-ROLE") Role role) {
 
-        return packageService.getVehicles(page, count, category);
+        if (role.equals(Role.ADMIN_PACKAGE) || role.equals(Role.USER))
+            return packageService.getVehicles(page, count, category);
+
+        throw new UnauthorizedException("Un authorized access to application");
 
     }
 
     @GetMapping(value = "/getHotels", params = {"page", "count", "category"}, produces = {MediaType.APPLICATION_JSON_VALUE})
     public ResponseUtil getHotelsByCategory(@RequestParam Integer page,
-                                              @RequestParam Integer count,
-                                              @RequestParam Integer category){
+                                            @RequestParam Integer count,
+                                            @RequestParam Integer category,
+                                            @RequestHeader("X-ROLE") Role role) {
 
-        return packageService.getHotels(page, count, category);
+        if (role.equals(Role.ADMIN_PACKAGE) || role.equals(Role.USER))
+            return packageService.getHotels(page, count, category);
 
+        throw new UnauthorizedException("Un authorized access to application");
     }
 
     @GetMapping(value = "/getHotelById", params = {"hotelId"}, produces = {MediaType.APPLICATION_JSON_VALUE})
-    public ResponseUtil getHotelsById(@RequestParam Integer hotelId){
+    public ResponseUtil getHotelsById(@RequestParam Integer hotelId, @RequestHeader("X-ROLE") Role role) {
 
-        return packageService.getHotelById(hotelId);
+        if (role.equals(Role.ADMIN_PACKAGE) || role.equals(Role.USER))
+            return packageService.getHotelById(hotelId);
 
+        throw new UnauthorizedException("Un authorized access to application");
     }
 
     @GetMapping(value = "/getVehicleById", params = {"vehicleId"}, produces = {MediaType.APPLICATION_JSON_VALUE})
-    public ResponseUtil getVehicleById(@RequestParam Integer vehicleId){
+    public ResponseUtil getVehicleById(@RequestParam Integer vehicleId, @RequestHeader("X-ROLE") Role role) {
 
-        return packageService.getVehicleById(vehicleId);
+        if (role.equals(Role.ADMIN_PACKAGE) || role.equals(Role.USER))
+            return packageService.getVehicleById(vehicleId);
 
+        throw new UnauthorizedException("Un authorized access to application");
     }
 
     @GetMapping(value = "/getGuideById", params = {"guideId"}, produces = {MediaType.APPLICATION_JSON_VALUE})
-    public ResponseUtil getGuideById(@RequestParam Integer guideId){
+    public ResponseUtil getGuideById(@RequestParam Integer guideId, @RequestHeader("X-ROLE") Role role) {
 
-        return packageService.getGuideById(guideId);
+        if (role.equals(Role.ADMIN_PACKAGE) || role.equals(Role.USER))
+            return packageService.getGuideById(guideId);
 
+        throw new UnauthorizedException("Un authorized access to application");
     }
 
     @GetMapping(value = "/getUsers", produces = {MediaType.APPLICATION_JSON_VALUE})
-    public ResponseUtil getAllUsers(){
+    public ResponseUtil getAllUsers(@RequestHeader("X-ROLE") Role role) {
+
+        if (!role.equals(Role.ADMIN_PACKAGE))
+            throw new UnauthorizedException("Un authorized access to application");
 
         return packageService.getAllUsers();
 
     }
 
     @GetMapping(value = "/getUserByNic", params = {"nic"}, produces = {MediaType.APPLICATION_JSON_VALUE})
-    public ResponseUtil getUserByNic(@RequestParam String nic){
+    public ResponseUtil getUserByNic(@RequestParam String nic, @RequestHeader("X-ROLE") Role role) {
+
+        if (!role.equals(Role.ADMIN_PACKAGE))
+            throw new UnauthorizedException("Un authorized access to application");
 
         return packageService.getUserByNic(nic);
 
     }
 
     @GetMapping(value = "/getNextPk")
-    public ResponseUtil getNextId(){
+    public ResponseUtil getNextId(@RequestHeader("X-ROLE") Role role) {
 
-        return ResponseUtil
-                .builder()
-                .code(200)
-                .data(packageService.getNextPk())
-                .message("Next Id get successful")
-                .build();
+        if (role.equals(Role.ADMIN_PACKAGE) || role.equals(Role.USER))
+            return ResponseUtil
+                    .builder()
+                    .code(200)
+                    .data(packageService.getNextPk())
+                    .message("Next Id get successful")
+                    .build();
+
+        throw new UnauthorizedException("Un authorized access to application");
     }
 
     @PostMapping(consumes = {MediaType.APPLICATION_JSON_VALUE}, produces = {MediaType.APPLICATION_JSON_VALUE})
-    public ResponseUtil savePackage(@RequestBody PackageDTO packageDTO){
+    public ResponseUtil savePackage(@RequestBody PackageDTO packageDTO, @RequestHeader("X-ROLE") Role role) {
 
-        return ResponseUtil
-                .builder()
-                .code(200)
-                .data(packageService.savePackage(packageDTO))
-                .message("Package save successful")
-                .build();
+        if (role.equals(Role.ADMIN_PACKAGE) || role.equals(Role.USER))
+            return ResponseUtil
+                    .builder()
+                    .code(200)
+                    .data(packageService.savePackage(packageDTO))
+                    .message("Package save successful")
+                    .build();
+
+        throw new UnauthorizedException("Un authorized access to application");
     }
 
     @PutMapping(consumes = {MediaType.APPLICATION_JSON_VALUE}, produces = {MediaType.APPLICATION_JSON_VALUE})
-    public ResponseUtil updatePackage(@RequestBody PackageDTO packageDTO){
+    public ResponseUtil updatePackage(@RequestBody PackageDTO packageDTO, @RequestHeader("X-ROLE") Role role) {
+
+        if (!role.equals(Role.ADMIN_PACKAGE))
+            throw new UnauthorizedException("Un authorized access to application");
 
         return ResponseUtil
                 .builder()
@@ -112,8 +139,11 @@ public class PackageController {
                 .build();
     }
 
-    @PutMapping(value = "/updateUserPackage",consumes = {MediaType.APPLICATION_JSON_VALUE}, produces = {MediaType.APPLICATION_JSON_VALUE})
-    public ResponseUtil updateUserPackage(@RequestBody PackageDTO packageDTO){
+    @PutMapping(value = "/updateUserPackage", consumes = {MediaType.APPLICATION_JSON_VALUE}, produces = {MediaType.APPLICATION_JSON_VALUE})
+    public ResponseUtil updateUserPackage(@RequestBody PackageDTO packageDTO, @RequestHeader("X-ROLE") Role role) {
+
+        if (!role.equals(Role.USER))
+            throw new UnauthorizedException("Un authorized access to application");
 
         return ResponseUtil
                 .builder()
@@ -124,7 +154,10 @@ public class PackageController {
     }
 
     @DeleteMapping(params = {"packageId"}, produces = {MediaType.APPLICATION_JSON_VALUE})
-    public ResponseUtil updatePackage(@RequestParam String packageId){
+    public ResponseUtil updatePackage(@RequestParam String packageId, @RequestHeader("X-ROLE") Role role) {
+
+        if (!role.equals(Role.ADMIN_PACKAGE))
+            throw new UnauthorizedException("Un authorized access to application");
 
         packageService.deletePackage(packageId);
         return ResponseUtil
@@ -135,7 +168,10 @@ public class PackageController {
     }
 
     @GetMapping(params = {"packageId"}, produces = {MediaType.APPLICATION_JSON_VALUE})
-    public ResponseUtil getById(@RequestParam String packageId){
+    public ResponseUtil getById(@RequestParam String packageId, @RequestHeader("X-ROLE") Role role) {
+
+        if (!role.equals(Role.ADMIN_PACKAGE))
+            throw new UnauthorizedException("Un authorized access to application");
 
         return ResponseUtil
                 .builder()
@@ -146,7 +182,12 @@ public class PackageController {
     }
 
     @GetMapping(value = "/getAll", params = {"page", "count"}, produces = {MediaType.APPLICATION_JSON_VALUE})
-    public ResponseUtil getPackagePageable(@RequestParam Integer page,@RequestParam Integer count){
+    public ResponseUtil getPackagePageable(@RequestParam Integer page,
+                                           @RequestParam Integer count,
+                                           @RequestHeader("X-ROLE") Role role) {
+
+        if (!role.equals(Role.ADMIN_PACKAGE))
+            throw new UnauthorizedException("Un authorized access to application");
 
         return ResponseUtil
                 .builder()
@@ -157,21 +198,30 @@ public class PackageController {
     }
 
     @GetMapping(value = "/getFreeGuide", params = {"startDate", "endDate"}, produces = {MediaType.APPLICATION_JSON_VALUE})
-    public ResponseUtil GetFreeGuide(@RequestParam String startDate, @RequestParam String endDate){
+    public ResponseUtil GetFreeGuide(@RequestParam String startDate,
+                                     @RequestParam String endDate,
+                                     @RequestHeader("X-ROLE") Role role) {
 
-        return ResponseUtil
-                .builder()
-                .code(200)
-                .data(packageService.getFreeGuides(startDate, endDate))
-                .message("Package getting all successful")
-                .build();
+        if (role.equals(Role.ADMIN_PACKAGE) || role.equals(Role.USER))
+            return ResponseUtil
+                    .builder()
+                    .code(200)
+                    .data(packageService.getFreeGuides(startDate, endDate))
+                    .message("Package getting all successful")
+                    .build();
+
+        throw new UnauthorizedException("Un authorized access to application");
     }
+
     @GetMapping(value = "/getPackageByNic", params = {"page", "count", "nic"}, produces = {MediaType.APPLICATION_JSON_VALUE})
     public ResponseUtil getPackagesByUserNic(@RequestParam Integer page,
                                              @RequestParam Integer count,
-                                             @RequestParam String nic){
+                                             @RequestParam String nic,
+                                             @RequestHeader("X-ROLE") Role role) {
 
-        System.out.println(page + count + nic);
+        if (!role.equals(Role.USER))
+            throw new UnauthorizedException("Un authorized access to application");
+
         return ResponseUtil
                 .builder()
                 .code(200)
@@ -180,11 +230,14 @@ public class PackageController {
                 .build();
     }
 
-    @PutMapping(value = "/payment",produces = {MediaType.APPLICATION_JSON_VALUE})
-    public ResponseUtil addPayment(@RequestParam("slip")MultipartFile file,
-                                   @RequestParam("packageId") String packageId){
+    @PutMapping(value = "/payment", produces = {MediaType.APPLICATION_JSON_VALUE})
+    public ResponseUtil addPayment(@RequestPart("slip") MultipartFile file,
+                                   @RequestPart("packageId") String packageId,
+                                   @RequestHeader("X-ROLE") Role role) {
 
-        System.out.println(file.getOriginalFilename());
+        if (!role.equals(Role.USER))
+            throw new UnauthorizedException("Un authorized access to application");
+
         packageService.addPaymentSlip(packageId, file);
         return ResponseUtil
                 .builder()
@@ -194,7 +247,10 @@ public class PackageController {
     }
 
     @PutMapping(value = "/confirm", params = "packageId", produces = {MediaType.APPLICATION_JSON_VALUE})
-    public ResponseUtil confirmPayment(@RequestParam String packageId){
+    public ResponseUtil confirmPayment(@RequestParam String packageId, @RequestHeader("X-ROLE") Role role) {
+
+        if (!role.equals(Role.ADMIN_PACKAGE))
+            throw new UnauthorizedException("Un authorized access to application");
 
         packageService.confirmPayment(packageId);
         return ResponseUtil
@@ -205,7 +261,10 @@ public class PackageController {
     }
 
     @PutMapping(value = "/reject", params = "packageId", produces = {MediaType.APPLICATION_JSON_VALUE})
-    public ResponseUtil rejectPayment(@RequestParam String packageId){
+    public ResponseUtil rejectPayment(@RequestParam String packageId, @RequestHeader("X-ROLE") Role role) {
+
+        if (!role.equals(Role.ADMIN_PACKAGE))
+            throw new UnauthorizedException("Un authorized access to application");
 
         packageService.rejectPayment(packageId);
         return ResponseUtil
@@ -216,8 +275,11 @@ public class PackageController {
     }
 
     @GetMapping(value = "/getPending", produces = {MediaType.APPLICATION_JSON_VALUE})
-    public ResponseUtil getPendingAllPayment(){
+    public ResponseUtil getPendingAllPayment(@RequestHeader("X-ROLE") Role role) {
 
+
+        if (!role.equals(Role.ADMIN_PACKAGE))
+            throw new UnauthorizedException("Un authorized access to application");
 
         return ResponseUtil
                 .builder()
@@ -227,8 +289,11 @@ public class PackageController {
                 .build();
     }
 
-    @GetMapping(value = "/getPendingOne", params =  "packageId",produces = {MediaType.APPLICATION_JSON_VALUE})
-    public ResponseUtil getPendingPayment(@RequestParam String packageId){
+    @GetMapping(value = "/getPendingOne", params = "packageId", produces = {MediaType.APPLICATION_JSON_VALUE})
+    public ResponseUtil getPendingPayment(@RequestParam String packageId, @RequestHeader("X-ROLE") Role role) {
+
+        if (!role.equals(Role.ADMIN_PACKAGE))
+            throw new UnauthorizedException("Un authorized access to application");
 
         return ResponseUtil
                 .builder()
@@ -239,8 +304,22 @@ public class PackageController {
     }
 
     @GetMapping(value = "/getAllPackages", produces = {MediaType.APPLICATION_JSON_VALUE})
-    public List<PackageDTO> getAllPackages(){
+    public List<PackageDTO> getAllPackages() {
 
         return packageService.getAllPackages();
+    }
+
+    @GetMapping(value = "/search", params = {"text"}, produces = {MediaType.APPLICATION_JSON_VALUE})
+    public ResponseUtil searchByText(@RequestParam String text, @RequestHeader("X-ROLE") Role role) {
+
+        if (!role.equals(Role.ADMIN_PACKAGE))
+            throw new UnauthorizedException("Un authorized access to application");
+
+        return ResponseUtil
+                .builder()
+                .code(200)
+                .message("Search package by text successfully !")
+                .data(packageService.searchByText(text))
+                .build();
     }
 }
